+++ src/main/java/um/tesoreria/core/hexagonal/mercadoPagoContext/domain/model/MercadoPagoContext.java
@@ -1,8 +1,10 @@
 package um.tesoreria.core.hexagonal.mercadoPagoContext.domain.model;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 import um.tesoreria.core.hexagonal.chequeraCuota.domain.model.ChequeraCuota;
 import um.tesoreria.core.hexagonal.umhub.reservaVacante.domain.model.ReservaVacante;
+import um.tesoreria.core.util.Jsonifier;
 
 import java.math.BigDecimal;
 import java.time.OffsetDateTime;
@@ -19,6 +21,7 @@ public class MercadoPagoContext {
     private Long chequeraCuotaId;
     private UUID reservaVacanteId;
     private String initPoint;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXX", timezone = "UTC")
     private OffsetDateTime fechaVencimiento;
 
     @Builder.Default
@@ -27,6 +30,7 @@ public class MercadoPagoContext {
     @Builder.Default
     private Byte changed = 0;
 
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXX", timezone = "UTC")
     private OffsetDateTime lastVencimientoUpdated;
     private String preferenceId;
     private String preference;
@@ -37,7 +41,9 @@ public class MercadoPagoContext {
     private Long chequeraPagoId;
     private String idMercadoPago;
     private String status;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXX", timezone = "UTC")
     private OffsetDateTime fechaPago;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXX", timezone = "UTC")
     private OffsetDateTime fechaAcreditacion;
 
     @Builder.Default
@@ -48,4 +54,7 @@ public class MercadoPagoContext {
     private ChequeraCuota chequeraCuota;
     private ReservaVacante reservaVacante;
 
+    public String jsonify() {
+        return Jsonifier.builder(this).build();
+    }
 }