+++ src/main/java/um/tesoreria/core/hexagonal/umhub/reservaVacante/domain/model/ReservaVacante.java
@@ -1,8 +1,10 @@
 package um.tesoreria.core.hexagonal.umhub.reservaVacante.domain.model;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 import um.tesoreria.core.hexagonal.domicilio.domain.model.Domicilio;
 import um.tesoreria.core.hexagonal.persona.domain.model.Persona;
+import um.tesoreria.core.hexagonal.umhub.campanha.domain.model.Campanha;
 import um.tesoreria.core.util.Jsonifier;
 
 import java.math.BigDecimal;
@@ -22,7 +24,9 @@ public class ReservaVacante {
     private UUID campanhaId;
     private String estado;
     private BigDecimal importe;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXX", timezone = "UTC")
     private OffsetDateTime vencimiento;
+    private Campanha campanha;
     private Persona persona;
     private Domicilio domicilio;
     private LocalDateTime created;