+++ src/main/java/um/tesoreria/core/model/dto/MercadoPagoContextDto.java
@@ -19,7 +19,7 @@ public class MercadoPagoContextDto {
     private Long chequeraCuotaId;
     private String initPoint;
 
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssZ", timezone = "UTC")
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXX", timezone = "UTC")
     private OffsetDateTime fechaVencimiento;
 
     @Builder.Default